Struct near_sdk_core::AttrSigInfo

pub struct AttrSigInfo {
    pub ident: Ident,
    pub non_bindgen_attrs: Vec<Attribute>,
    pub args: Vec<ArgInfo>,
    pub method_type: MethodType,
    pub is_payable: bool,
    pub is_private: bool,
    pub input_serializer: SerializerType,
    pub result_serializer: SerializerType,
    pub receiver: Option<Receiver>,
    pub returns: ReturnType,
    pub original_sig: Signature,
}

Information extracted from method attributes and signature.

Fields
ident: Ident

The name of the method.

non_bindgen_attrs: Vec<Attribute>

Attributes not related to bindgen.

args: Vec<ArgInfo>

All arguments of the method.

method_type: MethodType

Describes the type of the method.

is_payable: bool

Whether method accepting $NEAR.

is_private: bool

Whether method can accept calls from self (current account)

input_serializer: SerializerType

The serializer that we use for env::input().

result_serializer: SerializerType

The serializer that we use for the return type.

receiver: Option<Receiver>

The receiver, like mut self, self, &mut self, &self, or None.

returns: ReturnType

What this function returns.

original_sig: Signature

The original method signature.

Implementations

impl AttrSigInfo

pub fn input_struct(&self, input_struct_type: InputStructType) -> TokenStream2

Create struct representing input arguments.

input_struct_type represents whether the input structure will be used for serialization (e.g. for a promise input) or deserialization (e.g. for a method input). Each argument is getting converted to a field in a struct. Specifically argument: ATTRIBUTES ref mut binding @ SUBPATTERN : TYPE is getting converted to: binding: SUBTYPE, where TYPE is one of the following: & SUBTYPE, &mut SUBTYPE, SUBTYPE, and SUBTYPE is one of the following: [T; n], path like std::collections::HashMap<SUBTYPE, SUBTYPE>, or tuple (SUBTYPE0, SUBTYPE1, ...).

Example

struct Input {
  arg0: Vec<String>,
  arg1: [u64; 10],
  arg2: (u64, Vec<String>),
}

pub fn decomposition_pattern(&self) -> TokenStream2

Create pattern that decomposes input struct using correct mutability modifiers.

Example:

Input {
    arg0,
    mut arg1,
    arg2
}

pub fn constructor_expr(&self) -> TokenStream2

Create expression that constructs the struct.

Example:

Input {
    arg0,
    arg1,
    arg2,
}

pub fn arg_list(&self) -> TokenStream2

Create a sequence of arguments that can be used to call the method or the function of the smart contract.

Example:

a, &b, &mut c,

pub fn pat_type_list(&self) -> TokenStream2

Create a sequence of patterns and types to be used in the method signature.

Example:

a: u64, b: &mut T, ref mut c: Vec<String>,

pub fn callback_deserialization(&self) -> TokenStream2

Create code that deserializes arguments that were decorated with #[callback]

pub fn callback_vec_deserialization(&self) -> TokenStream2

Create code that deserializes arguments that were decorated with #[callback_vec].

impl AttrSigInfo

pub fn new(
original_attrs: &mut Vec<Attribute>,
original_sig: &mut Signature
) -> Result<Self>

Process the method and extract information important for near-sdk.

pub fn input_args(&self) -> impl Iterator<Item = &ArgInfo>

Only get args that correspond to env::input().