Struct near_sdk_core::ImplItemMethodInfo

pub struct ImplItemMethodInfo {
    pub attr_signature_info: AttrSigInfo,
    pub is_public: bool,
    pub struct_type: Type,
}

Information extracted from ImplItemMethod.

Fields
attr_signature_info: AttrSigInfo

Information on the attributes and the signature of the method.

is_public: bool

Whether method has pub modifier.

struct_type: Type

The type of the contract struct.

Implementations

impl ImplItemMethodInfo

pub fn method_wrapper(&self) -> TokenStream2

Generate wrapper method for the given method of the contract.

pub fn marshal_method(&self) -> TokenStream2

impl ImplItemMethodInfo

pub fn new(original: &mut ImplItemMethod, struct_type: Type) -> Result<Self>

Process the method and extract information important for near-sdk.

impl ImplItemMethodInfo

pub fn metadata_struct(&self) -> TokenStream2

Generates metadata struct for this method.

Example:

The following method:

fn f3(&mut self, arg0: FancyStruct, arg1: u64) -> Result<IsOk, Error> { }

will produce this struct:

near_sdk::MethodMetadata {
    name: "f3".to_string(),
    is_view: false,
    is_init: false,
    args: {
        #[derive(borsh::BorshSchema)]
        #[derive(serde :: Deserialize, serde :: Serialize)]
        struct Input {
            arg0: FancyStruct,
            arg1: u64,
        }
        Some(Input::schema_container())
    },
    callbacks: vec![],
    callbacks_vec: None,
    result: Some(Result < IsOk, Error > ::schema_container())
}

If args are serialized with Borsh it will not include #[derive(borsh::BorshSchema)].